refactor(board, eval): replace int16_t with int for mg_score and eg_score
- Update board_t struct to use int instead of int16_t for mg_score and eg_score
- Adjust set_piece, clear_piece, and static_eval to match new types

+++ src/board.h
@@ -15,8 +15,8 @@ typedef struct {
   bitboard_t occupancies[NR_OF_COLORS];
   bitboard_t occupancy;
   uint64_t zobrist;
-  int16_t mg_score[NR_OF_COLORS];
-  int16_t eg_score[NR_OF_COLORS];
+  int mg_score[NR_OF_COLORS];
+  int eg_score[NR_OF_COLORS];
   int num_moves;
   square_t kings[NR_OF_COLORS];
   uint8_t rights;
